requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
matplotlib>=3.7.0
shap>=0.45.0
joblib>=1.3.0
python-dotenv>=1.0.0

# Optional — app degrades gracefully without these
xgboost>=2.0.0
lightgbm>=4.0.0
wandb>=0.16.0
kagglehub>=0.2.0
